chore(suite-native): translations
